Split extension monoliths into ES modules now supported by browsers.
Extract background (settings, sorting, listeners, extract-domain) and popup
(render, events, constants) modules, and fix service worker init so listeners
register before cache warming to avoid popup hangs in E2E tests.
