+++ template-extension/popup/constants.js
@@ -1,0 +1,23 @@
+export const CHECKBOX_REVERSE = "ui_click_checkbox_sort_tabs_reverse";
+export const CHECKBOX_ALL_WINDOWS = "ui_click_checkbox_sort_tabs_all_windows";
+export const CHECKBOX_AUTO_ON_NEW_TAB =
+  "ui_click_checkbox_sort_tabs_auto_best_effort";
+export const CHECKBOX_RESPECT_TAB_GROUPS =
+  "ui_click_checkbox_sort_tabs_respect_tab_groups";
+export const CHECKBOX_SORT_PINNED = "ui_click_checkbox_sort_tabs_pinned";
+export const CHECKBOX_CLOSE_DUPLICATES =
+  "ui_click_checkbox_sort_tabs_close_duplicates";
+export const SELECT_DEFAULT_SORT_METHOD =
+  "ui_change_select_sort_select_tabs_default_sort_method";
+export const SELECT_SUSPENDED_TABS_POSITION =
+  "ui_change_select_suspended_tabs_position";
+export const SELECT_THEME = "ui_change_select_theme";
+
+export const COMMAND_DISPLAY_PRIORITY = {
+  command_sort_tabs_favicon_and_title: 1,
+  command_sort_tabs_title: 2,
+  command_sort_tabs_url: 3,
+  command_sort_tabs_mru: 4,
+  command_sort_tabs_shuffle: 5,
+  command_extract_domain: 6,
+};
